API

Important note

In contrast to the C-interface, we use one-based indices for the "high-level" API. Hence, the first NUMA node is NUMA node 1 (not 0). Consequently, on a system with 8 NUMA domains, numa_max_node() returns 8 (not 7).

Index

NUMA.current_cpu

NUMA.current_cpus

NUMA.current_numa_node

NUMA.current_numa_nodes

NUMA.ncpus

NUMA.nnumanodes

NUMA.numa_allocate_cpumask

NUMA.numa_allocate_nodemask

NUMA.numa_available

NUMA.numa_bitmask_alloc

NUMA.numa_bitmask_clearall!

NUMA.numa_bitmask_clearbit!

NUMA.numa_bitmask_setall!

NUMA.numa_bitmask_setbit!

NUMA.numa_get_membind

NUMA.numa_get_mems_allowed

NUMA.numa_max_node

NUMA.numa_max_possible_node

NUMA.numa_node_of_cpu

NUMA.numa_node_size

NUMA.numa_num_possible_nodes

NUMA.numa_pagesize

NUMA.numa_preferred

NUMA.numa_set_localalloc

NUMA.numa_set_membind

NUMA.numalocal

NUMA.numanode

NUMA.which_numa_node

NUMA.which_numa_node

NUMA.Bitmask

References - Add-ons

NUMA.current_cpu — Method

Returns the ID (starting at zero) of the CPU-thread on which the calling Julia thread is currently running on.

NUMA.current_cpus — Method

Returns the IDs (starting at zero) of the CPU-threads on which the Julia threads are currently running on.

NUMA.current_numa_node — Method

Returns the NUMA node associated with the CPU-thread on which the calling Julia thread is currently running on.

NUMA.current_numa_nodes — Method

Returns the NUMA nodes associated with the CPU-threads on which the Julia threads are currently running on.

NUMA.ncpus — Method

Returns the number CPU-threads available on this system.

NUMA.nnumanodes — Method

Returns the number of NUMA nodes available on this system.

NUMA.which_numa_node — Function

which_numa_node(arr::AbstractArray) -> Int64
which_numa_node(arr::AbstractArray, i; kwargs...) -> Int64

Query on which NUMA node the given array is allocated (assuming that it is allocated on a single node).

Technically only the address of a single element is checked (by default the first one). The optional second integer argument can be used to specify this element.

The optional keyword argument variant can be used to switch between methods used to determine the NUMA node. Valid options are :move_pages (default) and :get_mempolicy.

Note: Returned NUMA node IDs start at zero!

NUMA.which_numa_node — Method

which_numa_node(ptr::Ptr{T}; variant) -> Int64

Query on which NUMA node the page associated with the given memory address (pointer) is located.

The optional keyword argument variant can be used to switch between methods used to determine the NUMA node. Valid options are :move_pages (default) and :get_mempolicy.

Note: Returned NUMA node IDs start at zero!

NUMA.numalocal — Method

numalocal() -> NUMA.NUMALocal

Returns an array initializer that represents the local NUMA node. To be used as, e.g., Vector{Float64}(numalocal(), 1024).

NUMA.numanode — Method

numanode(i::Integer) -> NUMA.NUMANode

Returns an array initializer that represents a specific NUMA node. To be used as, e.g., Vector{Float64}(numanode(1), 1024).

References - Bitmask

NUMA.Bitmask — Type

Represents a bit mask. To be used to indicate, e.g., node masks or cpu masks.

Example:

julia> bm = Bitmask()
Bitmask (trunc): 00000000

julia> fill!(bm, 1); bm
Bitmask (trunc): 11111111

julia> bm[2] = 0; bm
Bitmask (trunc): 11111101

julia> numa_set_membind(bm)

julia> numa_get_membind()
Bitmask (trunc): 11111101

References

NUMA.numa_allocate_cpumask — Method

numa_allocate_cpumask() -> Bitmask

Returns a bitmask of a size equal to the kernel's cpu mask. In other words, large enough to represent all cpus.

NUMA.numa_allocate_nodemask — Method

numa_allocate_nodemask() -> Bitmask

Returns a bitmask of a size equal to the kernel's node mask. In other words, large enough to represent all nodes.

NUMA.numa_available — Method

numa_available() -> Bool

Is this a NUMA system?

NUMA.numa_bitmask_alloc — Function

numa_bitmask_alloc() -> Bitmask
numa_bitmask_alloc(n::Integer) -> Bitmask

Allocates a bitmask structure and its associated bit mask. The memory allocated for the bit mask contains enough words (type UInt64) to contain n bits. If n isn't provided, nnumanodes() is used.

The bit mask is zero-filled.

NUMA.numa_bitmask_clearall! — Method

numa_bitmask_clearall!(bm::Bitmask)

Set all bits in the given Bitmask to zero.

NUMA.numa_bitmask_clearbit! — Method

numa_bitmask_clearbit!(bm::Bitmask, n::Integer)

Clear the n-th bit of the given Bitmask.

NUMA.numa_bitmask_setall! — Method

numa_bitmask_setall!(bm::Bitmask)

Set all bits in the given Bitmask to one.

NUMA.numa_bitmask_setbit! — Method

numa_bitmask_setbit!(bm::Bitmask, n::Integer)

Set the n-th bit of the given Bitmask.

NUMA.numa_get_membind — Method

numa_get_membind() -> Bitmask

Returns the mask of nodes from which memory can currently be allocated

NUMA.numa_get_mems_allowed — Method

numa_get_mems_allowed() -> Bitmask

Returns the mask of nodes from which the process is allowed to allocate memory in it's current cpuset context

NUMA.numa_max_node — Method

numa_max_node() -> Int64

Highest node number available on the system

NUMA.numa_max_possible_node — Method

numa_max_possible_node() -> Int64

Number of the highest possible node in a system

NUMA.numa_node_of_cpu — Function

numa_node_of_cpu() -> Int64
numa_node_of_cpu(cpuid::Integer) -> Int64

Returns the NUMA node that the cpu with the given id (starting at zero) belongs to.

NUMA.numa_node_size — Function

numa_node_size(

) -> NamedTuple{(:memtot, :memfree), Tuple{Int64, Int64}}
numa_node_size(
    node::Integer
) -> NamedTuple{(:memtot, :memfree), Tuple{Int64, Int64}}

Returns a named tuple holding the total memory and free memory of the given NUMA node. If no node index is provided as an argument, current_numa_node() is used.

NUMA.numa_num_possible_nodes — Method

numa_num_possible_nodes() -> Int64

Returns the size of kernel's node mask, i.e. large enough to represent the maximum number of nodes that the kernel can handle

NUMA.numa_pagesize — Method

numa_pagesize() -> Int64

Returns the number of bytes in a page

NUMA.numa_preferred — Method

numa_preferred() -> Int64

Preferred NUMA node of the current task

NUMA.numa_set_localalloc — Method

numa_set_localalloc()

Sets the memory allocation policy for the calling task to local allocation

NUMA.numa_set_membind — Method

numa_set_membind(bm::Bitmask)

Sets the memory allocation mask. The task will only allocate memory from the nodes set in the given nodemask. Passing an empty nodemask or a nodemask that contains nodes other than those in the mask returned by numa_get_mems_allowed() will result in an error.